requirements.txt
asgiref==3.8.1
Django>=2.2
sqlparse==0.5.3
tzdata==2024.2
psycopg2-binary>=2.8
redis>=2.8
celery>=5.0
django-celery-beat>=2.6
python-binance==1.0.26
python-dotenv>=0.0.1
requests>=2.32
pandas>=2.2
pytz>=2024.2
django-environ>=0.11.2
pandas~=2.2         
pytest~=8.1
pytest-django~=4.8
coverage~=7.4
drf-spectacular~=0.27
drf-spectacular-sidecar~=2024.5



